==> 01_Limpeza_e_AnaliseInicial.ipynb <==
Ao fazer uma breve analise no csv, notei que algumas linhas tem algumas divergencias:
1. Comentarios faltando fechar aspas.
2. E as ultimas linhas apresentam colunas divergentes (Extremamente Negativo)

E essas caractaristicas foram tratas no util.cleaning,\
fechando o comentario e trocando o texto por ','\
tornando as colunas corretas por mais que com NaN

from utils.cleaning import fix_data_structure
from pathlib import Path

file_path = Path('data/01_original_data.csv')
fixed_file_path = fix_data_structure(file_path)

import pandas as pd

df = pd.read_csv('data/02_churn_fixed.csv')

df.columns = [
    'id',
    'name',
    'age',
    'gender',
    'location',
    'marital_status',
    'qtd_dependents',
    'start_date',
    'churn_date',
    'price',
    'service',
    'tma',
    'data_volume',
    'qtd_appointments',
    'comments'
]

df.tail(2) # tail pra validar que as ultimas linhas estao corretas

df['location'].value_counts()

Aqui vemos que temos sempre a capital de cada estado, então vou simplificar a notação apenas com no UF.

Obs.: Os dados apresentam uma incongruencia em: Rio de Janeiro/RJ/Rio Grande do Sul\
que deveria ser (provavelmente): Rio de Janeiro/RJ/Rio de Janeiro.\
Isso se resolve com a simplificação proposta

df['location'] = [splits[1] for splits in df['location'].str.split('/')]

df.tail(2)

Visto que temos bastante incongruencia nos dados, vamos aprofundar nas demais colunas

#### id

df['id'].value_counts()

df[df['id']==1068]

Parece que dois clientes estão dividindo o mesmo id, vamos verificar se esses clientes já existem baseado nos nomes

df[df['name']=='Julio Santos']

df[df['name']=='Gabriela Almeida']

Parece que não é o caso, vamos esolher um dos dois e adicionar um novo id

df['id'].describe()

df.loc[70, 'id'] = 1103

Validar se deu certo

df['id'].value_counts()

df['id'].describe()

#### name

df['name'].value_counts()

Alguns nome estão repetidos mas isso não é um problema por si só, vamos seguir assim mesmo

#### age

df['age'].value_counts()

Tudo certo

#### gender

df['gender'].value_counts()

Certo!

#### marital_status

df['marital_status'].value_counts()

Certo!

#### qtd_dependents

df['qtd_dependents'].value_counts() 

#### start_date

years = [date[0] for date in df['start_date'].str.split('-')]
months = [date[1] for date in df['start_date'].str.split('-')]
days = [date[2] for date in df['start_date'].str.split('-')]

print('years:', min(years), max(years))
print('months:', min(months), max(months))
print('days:', min(days), max(days))    

Os valores e estruturas estao coerentes com YYYY-MM-DD

#### churn_date

churn_dates = df['churn_date'].dropna()

years = [date[0] for date in churn_dates.str.split('-')]
months = [date[1] for date in churn_dates.str.split('-')]
days = [date[2] for date in churn_dates.str.split('-')]

print('years:', min(years), max(years))
print('months:', min(months), max(months))
print('days:', min(days), max(days))
print('churn_rate:', churn_dates.shape[0]/df.shape[0])

Os valores e estruturas estao coerentes com YYYY-MM-DD

Mas levanta aqui o ponto de que o churn percebido é muito baixo,\
a analise deve focar mais em detratores (baseado nos comentarios) do que no churn em si

#### price

df['price'].value_counts() 

Os valores e estruturas estao coerentes

#### service

df['service'].value_counts() 

Os valores e estruturas estao coerentes

#### tma

df['tma'].value_counts() 

Os valores estão coerentes, mas vamos ajustar de minutos para segundos

df['tma'] = df['tma'] * 60
df['tma'].fillna(0, inplace=True)

df['tma'].describe()

#### data_volume

df['data_volume'].value_counts() 

Vamos remover o GB para tratar o valor em si, se quisermos

try:
    df['data_volume'] = df['data_volume'].str.replace(' GB','')
    df['data_volume'] = df['data_volume'].str.replace('-','0')
    df['data_volume'] = df['data_volume'].astype(float)
except AttributeError:
    pass

df['data_volume'].fillna(0, inplace=True)
df['data_volume'].describe()

#### qtd_appointments

df['qtd_appointments'].value_counts() 

df['qtd_appointments'] = df['qtd_appointments'].str.replace('-','0')
df['qtd_appointments'] = df['qtd_appointments'].astype(float)
df['qtd_appointments'].value_counts() 

Vamos adicionar o atributo explicito de churn no df

df['churn'] = ~ df['churn_date'].isna()

E já vamos bater o olho nesses 4% de Churn

df.to_csv('data/03_churn_cleaned.csv', index=False)
