==> 02_Identificar_PromotorDetrator.ipynb <==
Vamos explorar os comentarios na ótica de Detrator vs Promotor para identificar valores de NPS\
Para isso vamos usar LLM (gpt-4o-mini) para aproveitar a caracteristica 'ready-to-use' e bem eficiente para essa classificação

Uma alternativa aqui poderia ser utilizar TF-IDF + Regressor, mas para isso dependeriamos de dados anotados,\
Talvez no futuro possamos coletar a nota de NPS junto já servindo como uma label e tornando esse processo mais rapido e barato em novas execuções

import time
from utils.nlp_classifier import classificar_comentario
import pandas as pd

df = pd.read_csv('data/03_churn_cleaned.csv')

labels = []
for comentario in df["comments"]:
    label = classificar_comentario(comentario)
    print(f"Comentário: {comentario}\n➡ Classificação: {label}\n")
    labels.append(label)
    time.sleep(1) # avoid rate limit

df["classificacao"] = labels
df.to_csv("data/04_nlp_classificados.csv", index=False)

Em validação manual os valores obtidos parecem bem coerentes

___

Vamos explorar como está nossa nota atual de NPS, considerando as classificações do LLM

Baseado na formula tradicional de NPS:

$NPS = \frac{\text{Promotores}}{\text{Total}} - \frac{\text{Detratores}}{\text{Total}}$

Que varia de -100 a 100
E tem algumas zonas de interesse:
| Faixa de NPS  | Classificação           | Interpretação comum                   |
| ------------- | ----------------------- | ------------------------------------- |
| **90 a 100**  | Zona de excelência      | Lealdade excepcional                  |
| **70 a 89**   | Zona de qualidade       | Altamente recomendável                |
| **50 a 69**   | Zona de sucesso         | Muito bom, mas pode melhorar          |
| **0 a 49**    | Zona de aperfeiçoamento | Ok, mas ainda há espaço para melhorar |
| **-1 a -100** | Zona crítica            | Urgente investigar problemas          |


import pandas as pd
from utils.nlp_classifier import calcular_nps

df = pd.read_csv('data/04_nlp_classificados.csv')

nps_geral = calcular_nps(df, verbose=True)

O NPS atual descreve .98, praticamente 1 ponto, o que nos coloca na beira da Zona Critica, mas ainda na Zona de Aperfeiçoamento

### Explorando NPS de sub-grupos

Percebo 4 atributos que podem ter relação com a satisfação do cliente:
1. location (já que podemos ter localidades com qualidades diferentes de serviço)
2. service (já que podemos ter serviços com qualidades diferentes)
3. qtd_dependets (já que isso impacta diretamente em como os serviços sao consumidos)
4. price (já que eu posso tanto ter clientes insatisfeitos com os serviços mais basicos,\
          como clientes mais exigentes com os serviços mais caros)

Vamos explorar como é o NPS em cada um dos subgrupos desses atributos

for col in ['location', 'service', 'qtd_dependents', 'price']:
    print(f"\nNPS por '{col}':")
    for value in df[col].unique():
        temp_df = df[df[col]==value]
        nps = calcular_nps(temp_df, verbose=False)
        print(f"    {value}: {nps}")


Todos os valores de location, service, qtd_dependets flutuam nas duas faixas mais baixas: Zona Critica e Zona de Aprimoramento.

Tendo uma tensão um pouco maior em SP.\
E um pouco melhor em Telefonia Fixa.

Quanto a hipotese de que maior numero de dependentes impactar negativamente na satisfação do cliente:\
isso não é sustentado pelos valores de NPS nesses subgrupos

Em especial ao olhar para preço temos uma flutuação bem grande que provavelmente se deve ao tamanho da base\
O que faz com que a quantidade de cada classe do NPS seja bem pequena, podemos tornar essa analise mais robusta utilizando faixas de preço\
ao inves do preço em si, caso essa informação pareça relevante mais pra frente

Nos proximos IPYNB vamos explorar ações que podem melhorar esses valores
