==> 03_Identificar_Topicos.ipynb <==
Vamos explorar um pouco os principais topicos de apresentados nos comentarios de cada usuario.

Para esse fim poderiamos utilizar uma abordagem baseada em LLM como na Identificação de Promotor e Detrator.\
Mas na intenção de demonstrar um maior leque de conhecimentos e embassamento classico, vamos utilizar BERTopic\
Que demanda do conhecimento de preprocessamento e limpeza classicos de NLP

import pandas as pd

df = pd.read_csv('data/04_nlp_classificados.csv')

import re
import unidecode
import nltk
from nltk.corpus import stopwords
from nltk.stem import RSLPStemmer

# Baixar stopwords e stemmer do NLTK (roda só uma vez)
nltk.download("stopwords")
nltk.download('rslp')

# Lista de stopwords em português
stop_words = set(stopwords.words("portuguese"))

stemmer = RSLPStemmer()

# Função de pré-processamento
def preprocess(text):
    text = unidecode.unidecode(text)
    text = text.lower()
    text = re.sub(r"[^\w\s]", "", text)
    text = re.sub(r"\d+", "", text)
    tokens = text.split()
    tokens = [
        stemmer.stem(t) 
        for t in tokens 
        if t not in stop_words and len(t) > 4
    ]
    return " ".join(tokens)

# Pré-processar os comentários
df["comments_pre"] = df["comments"].astype(str).apply(preprocess)

Acima usamos nltk para pegar uma lista pronta de stopwords\
(que descrevem palavras sem conteudo proprio, apenas apoiam a coerencia do texto).


from bertopic import BERTopic
from sentence_transformers import SentenceTransformer

# Embedding model (CPU)
embedding_model = SentenceTransformer("paraphrase-multilingual-MiniLM-L12-v2", device="cpu")

# Criar modelo BERTopic
topic_model = BERTopic(embedding_model=embedding_model, 
                        language="portuguese",
                        min_topic_size=6)

# Gerar tópicos
topics, probs = topic_model.fit_transform(df["comments_pre"])

# Adicionar ao DataFrame
df["topic"] = topics

df.drop(columns=["comments_pre"], inplace=True)

# Mostrar frequência dos tópicos
print(df["topic"].value_counts())

Foram encontrados 6 topicos (e 1 topico nulo) ao longo dos comentarios

topic_model.get_topic_info()

O BERTopic apresenta alguns nomes baseado nas palavras chaves, mas para facilitar o entendimento podemos renomear:

map_topic = {
    -1:"other",
    0:"internet_connection",
    1:"critical_appointments",
    2:"quality_tv_channels",
    3:"signal"
    }

df["topic"] = df["topic"].map(map_topic)

Em analise ocular encontrei mais um topico que acabou não sendo descrito pelo BERTopic: Atendimento

vamos adiciona-lo manualmente

# Definir tópico 'atendimento' onde 'atendimento' aparecer nos comentários
df.loc[df['comments'].str.contains('atendimento', case=False, na=False), 'topic'] = 'atendimento'

Aqui de forma bem simples, pra todo comentario que tiver a palavra 'atendimento' assumimos que o assunto é 'atendimento'

Explorei usar o lemma 'atend' mas muitos comentarios utilizam expressões como 'o serviço tem atendido minhas expectativas'

df.to_csv("data/05_nlp_classificados_topicos.csv", index=False)

Os topicos que conseguimos parece sempre ter uma relação muito grande com o serviço, vamos investigar essa relação para entender a utilidade desses topicos

import pandas as pd

df = pd.read_csv("data/05_nlp_classificados_topicos.csv")

df.head(2)

for service in df["service"].unique():
    print(service)
    for topic in df["topic"].unique():
        temp_df = df[(df["service"]==service) & (df["topic"]==topic)]
        print(f"    {topic}: {temp_df.shape[0]}")


Basicamente com exceção de critical_appointments e atendimento, cada outro topico é bem associado a um serviço

Vamos explorar especificamente os detratores

df_detratores = df[df["classificacao"]=="Detrator"]

for service in df_detratores["service"].unique():
    print(service)
    for topic in df_detratores["topic"].unique():
        temp_df = df_detratores[(df_detratores["service"]==service) & (df_detratores["topic"]==topic)]
        print(f"    {topic}: {temp_df.shape[0]}")


Certo, aqui podemos começar a ter insights sobre onde atuar:

**critical_appointments** no geral tem pouca informação, é apenas alguem bem chateado com os serviços insultando a empresa.\
Talvez possamos encontrar nesses casos reclamações anteriores que possam indicar melhor as dores do cliente e possiveis caminhos para reverter esse, provavel, churn eminente

Logo depois tempos **Atendimento**, que chama bastante atençao por ser um topico cross os serviços.

E então temos as reclamações relacionadas ao serviço em si: **internet_connection** e **signal**

Vamos explorar uma ultima ótica deste ultimos 2 topicos no proximo IPYNB
